AppCriptografia/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.17
idna==3.10
mido==1.3.3
mysqlclient==2.2.6
music21==8.3.0
Naked==0.1.32
pycryptodome==3.21.0
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
numpy==2.0.2
networkx==3.2.1
pygraphviz==1.11